fix: g2plot render wrong when fields contains `.`
