+++ packages/plugins/data-visualization/src/client/chart/antd/antd.ts
@@ -1,0 +1,10 @@
+import { Chart } from '../chart';
+
+export class AntdChart extends Chart {
+  getReference() {
+    return {
+      title: this.title,
+      link: `https://ant.design/components/${this.name}`,
+    };
+  }
+}
