+++ packages/plugins/data-visualization/src/client/chart/antd/index.ts
@@ -1,0 +1,4 @@
+import { Statistic } from './statistic';
+import { Table } from './table';
+
+export default [new Statistic(), new Table()];
