+++ packages/plugins/data-visualization/src/client/index.tsx
@@ -2,10 +2,12 @@ import { Plugin, SchemaComponentOptions, SchemaInitializerContext, SchemaInitial
 import React, { useContext } from 'react';
 import { ChartInitializers, ChartV2Block, ChartV2BlockDesigner, ChartV2BlockInitializer } from './block';
 import { useChartsTranslation } from './locale';
-import { ChartRenderer, ChartRendererProvider, InternalLibrary } from './renderer';
-import { ChartLibraryProvider } from './renderer/ChartLibrary';
+import { ChartRenderer, ChartRendererProvider } from './renderer';
+import { ChartLibraryProvider } from './chart/library';
+import g2plot from './chart/g2plot';
+import antd from './chart/antd';
 
-const Chart: React.FC = (props) => {
+const DataVisualization: React.FC = (props) => {
   const { t } = useChartsTranslation();
   const initializers = useContext<any>(SchemaInitializerContext);
   const children = initializers.BlockInitializers.items[0].children;
@@ -29,7 +31,7 @@ const Chart: React.FC = (props) => {
       }}
     >
       <SchemaInitializerProvider initializers={{ ...initializers, ChartInitializers }}>
-        <ChartLibraryProvider name="Built-in" charts={InternalLibrary}>
+        <ChartLibraryProvider name="Built-in" charts={[...g2plot, ...antd]}>
           {props.children}
         </ChartLibraryProvider>
       </SchemaInitializerProvider>
@@ -39,9 +41,11 @@ const Chart: React.FC = (props) => {
 
 class DataVisualizationPlugin extends Plugin {
   async load() {
-    this.app.addProvider(Chart);
+    this.app.addProvider(DataVisualization);
   }
 }
 
 export default DataVisualizationPlugin;
 export { ChartLibraryProvider };
+export { Chart } from './chart/chart';
+export type { ChartType } from './chart/chart';
