+++ packages/plugins/data-visualization/src/client/renderer/index.ts
@@ -1,4 +1,2 @@
-export * from './ChartLibrary';
 export * from './ChartRenderer';
 export * from './ChartRendererProvider';
-export * from './library';
