--- packages/plugins/data-visualization/src/client/renderer/library/index.tsx
@@ -1,4 +1,0 @@
-import { AntdLibrary } from './AntdLibrary';
-import { G2PlotLibrary } from './G2PlotLibrary';
-
-export const InternalLibrary = { ...G2PlotLibrary, ...AntdLibrary };
